+++ include/util/iutest_util_no_yoda.hpp
@@ -6,7 +6,7 @@
  *
  * @author      t.shirayanagi
  * @par         copyright
- * Copyright (C) 2015-2016, Takazumi Shirayanagi\n
+ * Copyright (C) 2015-2018, Takazumi Shirayanagi\n
  * This software is released under the new BSD License,
  * see LICENSE
 */
@@ -68,6 +68,22 @@
 #  define   IUTEST_ASSUME_DOUBLE_EQ(actual, expected)       IUTEST_TEST_DOUBLE_EQ(expected, actual, IUTEST_ASSUME_FAILURE)
 #endif
 
+#ifdef IUTEST_ASSERT_LONG_DOUBLE_EQ
+#  undef IUTEST_ASSERT_LONG_DOUBLE_EQ
+#  define   IUTEST_ASSERT_LONG_DOUBLE_EQ(actual, expected)  IUTEST_TEST_LONG_DOUBLE_EQ(expected, actual, IUTEST_ASSERT_FAILURE)
+#endif
+#ifdef IUTEST_EXPECT_LONG_DOUBLE_EQ
+#  undef IUTEST_EXPECT_LONG_DOUBLE_EQ
+#  define   IUTEST_EXPECT_LONG_DOUBLE_EQ(actual, expected)  IUTEST_TEST_LONG_DOUBLE_EQ(expected, actual, IUTEST_EXPECT_FAILURE)
+#endif
+#ifdef IUTEST_INFORM_LONG_DOUBLE_EQ
+#  undef IUTEST_INFORM_LONG_DOUBLE_EQ
+#  define   IUTEST_INFORM_LONG_DOUBLE_EQ(actual, expected)  IUTEST_TEST_LONG_DOUBLE_EQ(expected, actual, IUTEST_INFORM_FAILURE)
+#endif
+#ifdef IUTEST_ASSUME_LONG_DOUBLE_EQ
+#  undef IUTEST_ASSUME_LONG_DOUBLE_EQ
+#  define   IUTEST_ASSUME_LONG_DOUBLE_EQ(actual, expected)  IUTEST_TEST_LONG_DOUBLE_EQ(expected, actual, IUTEST_ASSUME_FAILURE)
+#endif
 
 #ifdef IUTEST_ASSERT_STREQ
 #  undef IUTEST_ASSERT_STREQ