+++ test/matcher_tests.cpp
@@ -41,6 +41,10 @@ int* p1 = NULL;
 int** p2 = &p1;
 float f0 = 0.0f;
 double d0 = 0.0;
+#if IUTEST_HAS_LONG_DOUBLE
+long double ld0 = 0.0;
+#endif
+
 struct X { int a, b; X(int _a, int _b) : a(_a), b(_b) {} int GetA() const { return a; } };
 X gx(1, 1);
 ::std::map<int, X> mx;
@@ -100,24 +104,95 @@ IUTEST(Matcher, TypedEq)
 IUTEST(Matcher, FloatEq)
 {
     IUTEST_EXPECT_THAT(f0, FloatEq(0.0f));
+    IUTEST_EXPECT_THAT(f0, FloatingPointEq(0.0f));
 }
 
 IUTEST(Matcher, DoubleEq)
 {
     IUTEST_EXPECT_THAT(d0, DoubleEq(0.0));
+    IUTEST_EXPECT_THAT(d0, FloatingPointEq(0.0));
 }
+
+IUTEST(Matcher, FloatNear)
+{
+    IUTEST_EXPECT_THAT(f0, FloatNear(0.0f, 0.0f));
+    IUTEST_EXPECT_THAT(f0, FloatNear(1.0f, 1.0f));
+    IUTEST_EXPECT_THAT(f0, FloatingPointNear(0.0f, 0.0f));
+}
+
+IUTEST(Matcher, DoubleNear)
+{
+    IUTEST_EXPECT_THAT(d0, DoubleNear(0.0, 0.0));
+    IUTEST_EXPECT_THAT(d0, DoubleNear(1.0, 1.0));
+    IUTEST_EXPECT_THAT(d0, FloatingPointNear(0.0, 0.0));
+}
+
+#if IUTEST_HAS_LONG_DOUBLE
+
+IUTEST(Matcher, LongDoubleEq)
+{
+    IUTEST_EXPECT_THAT(ld0, LongDoubleEq(0.0));
+    long double ldx = 0.001;
+    IUTEST_EXPECT_THAT(ldx, LongDoubleEq(0.001));
+}
+
+IUTEST(Matcher, LongDoubleNear)
+{
+    IUTEST_EXPECT_THAT(ld0, LongDoubleNear(0.0, 0.0));
+    long double ldx = 0.001;
+    IUTEST_EXPECT_THAT(ld0, LongDoubleNear(ldx, ldx));
+}
+#endif
 
 IUTEST(Matcher, NanSensitiveFloatEq)
 {
     IUTEST_EXPECT_THAT(f0, NanSensitiveFloatEq(0.0f));
     IUTEST_EXPECT_THAT(0.0f/f0, NanSensitiveFloatEq(0.0f/f0));
+    IUTEST_EXPECT_THAT(0.0f/f0, NanSensitiveFloatingPointEq(0.0f/f0));
 }
 
 IUTEST(Matcher, NanSensitiveDoubleEq)
 {
     IUTEST_EXPECT_THAT(d0, NanSensitiveDoubleEq(0.0));
     IUTEST_EXPECT_THAT(0.0/d0, NanSensitiveDoubleEq(0.0/d0));
+    IUTEST_EXPECT_THAT(0.0/d0, NanSensitiveFloatingPointEq(0.0/d0));
 }
+
+IUTEST(Matcher, NanSensitiveFloatNear)
+{
+    IUTEST_EXPECT_THAT(f0, NanSensitiveFloatNear(0.0f, 0.0f));
+    IUTEST_EXPECT_THAT(f0, NanSensitiveFloatNear(1.0f, 1.0f));
+    IUTEST_EXPECT_THAT(0.0f/f0, NanSensitiveFloatNear(0.0f/f0, 0.0f));
+    IUTEST_EXPECT_THAT(0.0f/f0, NanSensitiveFloatingPointNear(0.0f/f0, 0.0f));
+}
+
+IUTEST(Matcher, NanSensitiveDoubleNear)
+{
+    IUTEST_EXPECT_THAT(d0, NanSensitiveDoubleNear(0.0, 0.0));
+    IUTEST_EXPECT_THAT(d0, NanSensitiveDoubleNear(1.0, 1.0));
+    IUTEST_EXPECT_THAT(0.0/d0, NanSensitiveDoubleNear(0.0/d0, 0.0));
+    IUTEST_EXPECT_THAT(0.0/d0, NanSensitiveFloatingPointNear(0.0/d0, 0.0));
+}
+
+#if IUTEST_HAS_LONG_DOUBLE
+
+IUTEST(Matcher, NanSensitiveLongDoubleEq)
+{
+    IUTEST_EXPECT_THAT(ld0, NanSensitiveLongDoubleEq(0.0));
+    IUTEST_EXPECT_THAT(0.0/ld0, NanSensitiveLongDoubleEq(0.0/ld0));
+    long double ldx = 0.001;
+    IUTEST_EXPECT_THAT(ldx, NanSensitiveLongDoubleEq(0.001));
+}
+
+IUTEST(Matcher, NanSensitiveLongDoubleNear)
+{
+    IUTEST_EXPECT_THAT(ld0, NanSensitiveLongDoubleNear(0.0, 0.0));
+    IUTEST_EXPECT_THAT(0.0/ld0, NanSensitiveLongDoubleNear(0.0/ld0, 0.0));
+    long double ldx = 0.001;
+    IUTEST_EXPECT_THAT(ld0, NanSensitiveLongDoubleNear(ldx, ldx));
+}
+
+#endif
 
 IUTEST(Matcher, StrEq)
 {
@@ -377,23 +452,46 @@ IUTEST(MatcherFailure, FloatEq)
 {
     CHECK_FAILURE( IUTEST_ASSERT_THAT(f0, FloatEq(1.0f)), "Eq: " );
     CHECK_FAILURE( IUTEST_ASSERT_THAT(0/f0, FloatEq(0/f0)), "Eq: " );
+    CHECK_FAILURE( IUTEST_ASSERT_THAT(0/f0, FloatingPointEq(0/f0)), "Eq: " );
 }
 
 IUTEST(MatcherFailure, DoubleEq)
 {
     CHECK_FAILURE( IUTEST_ASSERT_THAT(d0, DoubleEq(1.0)), "Eq: " );
     CHECK_FAILURE( IUTEST_ASSERT_THAT(0/d0, DoubleEq(0/d0)), "Eq: " );
+    CHECK_FAILURE( IUTEST_ASSERT_THAT(0/d0, FloatingPointEq(0/d0)), "Eq: " );
 }
+
+#if IUTEST_HAS_LONG_DOUBLE
+
+IUTEST(MatcherFailure, LongDoubleEq)
+{
+    CHECK_FAILURE( IUTEST_ASSERT_THAT(ld0, LongDoubleEq(1.0)), "Eq: " );
+    CHECK_FAILURE( IUTEST_ASSERT_THAT(0/ld0, LongDoubleEq(0/ld0)), "Eq: " );
+}
+
+#endif
 
 IUTEST(MatcherFailure, NanSensitiveFloatEq)
 {
     CHECK_FAILURE( IUTEST_ASSERT_THAT(f0, NanSensitiveFloatEq(1.0f)), "Eq: " );
+    CHECK_FAILURE( IUTEST_ASSERT_THAT(f0, NanSensitiveFloatingPointEq(1.0f)), "Eq: " );
 }
 
 IUTEST(MatcherFailure, NanSensitiveDoubleEq)
 {
     CHECK_FAILURE( IUTEST_ASSERT_THAT(d0, NanSensitiveDoubleEq(1.0)), "Eq: " );
+    CHECK_FAILURE( IUTEST_ASSERT_THAT(d0, NanSensitiveFloatingPointEq(1.0)), "Eq: " );
 }
+
+#if IUTEST_HAS_LONG_DOUBLE
+
+IUTEST(MatcherFailure, NanSensitiveLongDoubleEq)
+{
+    CHECK_FAILURE( IUTEST_ASSERT_THAT(ld0, NanSensitiveLongDoubleEq(1.0)), "Eq: " );
+}
+
+#endif
 
 IUTEST(MatcherFailure, StrEq)
 {