+++ CHANGES.md
@@ -4,6 +4,11 @@
 
 ## Changes for 1.16.6
 
+* Improved
+  * Matcher に (NanSensitive)FloatNear/DoubleNear を追加
+  * Matcher に template 版の浮動小数点数比較を追加・(NanSensitive)FloatingPointEq/FloatingPointNear
+  * long double に対応
+  
 * Changes
   * iuwandbox: -D__WANDBOX__ コンパイラーオプションを必ずつけるように変更